extracted product storer into a separate file
